pyproject.toml
[project]
name = "01-16-2026-policy-catalogue-automation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "feedparser>=6.0.12",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "gitpython>=3.1.43",
    "lxml>=5.1.0",
]
